web/requirements.txt
Flask==3.0.0
Flask-RESTful==0.3.10
pymongo==4.6.1
bcrypt==4.1.2
numpy==1.26.3
Keras
requests==2.31.0
tensorflow # Specify a compatible version
Pillow==8.3.2
